[package_server rename] Fix more includes
