+++ debian/pulse2-package-server.install
@@ -1,6 +1,6 @@
 usr/sbin/pulse2-package-server
 usr/sbin/pulse2-package-server-register-imaging
-usr/lib/python2.?/*-packages/pulse2/package_server
+usr/lib/python2.?/*-packages/pulse2/package_server2
 etc/mmc/pulse2/package-server
 etc/mmc/pulse2/atftpd/pcre.conf
 var/lib/pulse2/imaging
